Ignore ipap11helper/setup.py in doctests
Pytest imports all modules when running doctests.
The setup.py runs code on import, and raises an exception,
depending on globa connand-line arguments, so it needs to be ignored.

Also, pytest dislikes multiple top-level modules with the same name
("setup" in this case). Again ignoring is the way to go.