Remove the removal of the ccache
It is not necessary to remove the ccache on upgrades on modern IPA
servers, even if the ccache contains stale data either it is re-initialized by
mod_auth_kerb or a new ccache collection is created (if completely unrelated
credentials were present), at least when using DIR or keyring ccaches.

This line causes wrong SELinux labels to be set in the kernel keyring on
uprades, which the cause the apache server to fail to use th ccache.

https://fedorahosted.org/freeipa/ticket/4815

Reviewed-By: Martin Kosek <mkosek@redhat.com>