Rename caught exception for use outside the except: block.
In Python 3, the variable with the currently handled exception is unset
at the end of the except block. (FWIW, this is done to break reference
cycles, since exception instances now carry tracebacks, which contain
all locals.)

Use a simpler structure for the ipatests.raises utility that only uses the
exception inside the except block.

Also fix a similar situation in baseldap's error handler.